Move Person admin and helpers to People app
