+++ ynr/apps/popolo/admin.py
@@ -10,53 +10,6 @@ from .behaviors import admin as generics
 from django.utils.translation import ugettext_lazy as _
 
 
-class MembershipInline(admin.StackedInline):
-    extra = 0
-    model = models.Membership
-
-
-class PersonAdmin(admin.ModelAdmin):
-    fieldsets = (
-        (None, {"fields": ("name", "gender", "birth_date", "death_date")}),
-        (
-            "Biography",
-            {
-                "classes": ("collapse",),
-                "fields": ("summary", "image", "biography"),
-            },
-        ),
-        (
-            "Honorifics",
-            {
-                "classes": ("collapse",),
-                "fields": ("honorific_prefix", "honorific_suffix"),
-            },
-        ),
-        (
-            "Demography",
-            {"classes": ("collapse",), "fields": ("national_identity",)},
-        ),
-        (
-            "Special Names",
-            {
-                "classes": ("collapse",),
-                "fields": (
-                    "family_name",
-                    "given_name",
-                    "additional_name",
-                    "patronymic_name",
-                    "sort_name",
-                ),
-            },
-        ),
-        (
-            "Advanced options",
-            {"classes": ("collapse",), "fields": ("start_date", "end_date")},
-        ),
-    )
-    inlines = generics.BASE_INLINES + [MembershipInline]
-
-
 class PostAdmin(admin.ModelAdmin):
     model = models.Post
     fieldsets = (
@@ -98,5 +51,4 @@ class OrganizationAdmin(admin.ModelAdmin):
 
 
 admin.site.register(models.Post, PostAdmin)
-admin.site.register(models.Person, PersonAdmin)
 admin.site.register(models.Organization, OrganizationAdmin)
